requirements.txt
biopython>=1.78
requests>=2.25.1
argparse